MxOnline-bobby/requirements.txt
Django==1.9
django-crispy-forms==1.6.1
django-formtools==1.0
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
DjangoUeditor==1.8.143
httplib2==0.9.2
Pillow==3.4.2

